requirements.txt
booktype
python-telegram-bot
PyPDF2
pdfplumber
python-dotenv
psutil
redis
aiosqlite
cachetools